requirements.txt
streamlit==1.38.0
altair<5
tqdm
langchain
langchain_openai
langchain_community
faiss-cpu
python-dotenv
